requirements.txt
google-api-python-client
google_auth_oauthlib
djangorestframework
django
# django-channels
channels
daphne
dotenv